+++ src/main/java/com/aoe/restapi/utility/facade/UtilFacade.java
@@ -1,10 +1,19 @@
 package com.aoe.restapi.utility.facade;
 
+import com.aoe.restapi.exception.exception.ArgException;
+import com.aoe.restapi.exception.exception.AuthenticationException;
 import com.aoe.restapi.exception.exception.JwtNotValidException;
+import com.aoe.restapi.model.entity.User;
+import com.aoe.restapi.model.entity.UserRole;
 import com.aoe.restapi.utility.auth.JwtUtil;
 import com.aoe.restapi.utility.http.HttpUtil;
+import org.aspectj.lang.JoinPoint;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.Arrays;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 @Service
 public class UtilFacade {
@@ -16,17 +25,12 @@ public class UtilFacade {
         this.jwtUtil = jwtUtil;
     }
 
-    public int performJwtOperations() {
+    public int getIdFromJwtHeader() {
         // get token from header
         String jwtToken = HttpUtil.getTokenFromHeader();
 
-        // validate jwt token
-        if (!jwtUtil.validateToken(jwtToken))
-            throw new JwtNotValidException();
-
         // get id from jwt
-        int userId = Integer.parseInt(jwtUtil.getIdFromToken(jwtToken));
-        return userId;
+        return Integer.parseInt(jwtUtil.getIdFromToken(jwtToken));
     }
 
     public void commonJwtAuthentication() {
@@ -36,5 +40,33 @@ public class UtilFacade {
         // validate jwt token
         if (!jwtUtil.validateToken(jwtToken))
             throw new JwtNotValidException();
+    }
+
+
+    public Object[] getArguments(JoinPoint joinPoint) {
+        // get the arguments
+        Object[] args = joinPoint.getArgs();
+
+        // check arguments
+        if (args.length == 0)
+            throw new ArgException();
+
+        return args;
+    }
+
+    public static void checkAuthentication(User user, String[] roles) {
+        if (!user.getRoleSet().stream().anyMatch(role -> role.getName().equals("project manager")))
+            throw new AuthenticationException();
+
+        Set<String> userRoles = user.getRoleSet().stream()
+                .map(UserRole::getName)
+                .collect(Collectors.toSet());
+
+        boolean hasAnyRole = Arrays.stream(roles)
+                .anyMatch(userRoles::contains);
+
+        if (!hasAnyRole) {
+            throw new AuthenticationException();
+        }
     }
 }
